Move shard listing methods to a new class
